requirements.txt
streamlit==1.12.0
google-auth==2.6.2
google-auth-oauthlib==0.5.1
google-auth-httplib2==0.1.0
google-api-python-client==2.48.0
openai==0.19.0
python-dotenv==0.20.0